Refactor BrainGlobe related modules into new sub-package
- Group into a sub-package
- Rename the modules to reflect the Model-View-Controller organization
